// src/app/core/api/generated/model/attachment-dto.ts
/**
 * B12ForK12 V3 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AttachmentDto { 
    fileName?: string;
    fileUploadName?: string;
    fileSizeInBytes: number;
    fileURL?: string;
    fileThumbnailURL?: string;
    fileType: number;
    fileExtension?: string;
    fileContentType?: string;
    originalFileUrl?: string;
}