// src/app/core/api/generated/model/recipient-survey-condition-dto.ts
/**
 * B12ForK12 V3 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RecipientSurveyConditionDto { 
    dependsOnQuestionInstanceId: string;
    triggerChoiceIds?: Array<string>;
}