// src/app/core/api/generated/model/recipient-survey-distribution-dto.ts
/**
 * B12ForK12 V3 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecipientSurveySectionDto } from './recipient-survey-section-dto';


export interface RecipientSurveyDistributionDto { 
    distributionId: string;
    title?: string;
    introHtml?: string;
    startUtc: string;
    endUtc: string;
    oneAttemptNoEdit: boolean;
    isSubmitted: boolean;
    canEditSubmission: boolean;
    submittedAtUtc?: string;
    sections?: Array<RecipientSurveySectionDto>;
}