// src/app/core/api/generated/model/recipient-survey-existing-answer-dto.ts
/**
 * B12ForK12 V3 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AttachmentDto } from './attachment-dto';


export interface RecipientSurveyExistingAnswerDto { 
    ratingValue?: number;
    textValue?: string;
    selectedChoiceIds?: Array<string>;
    attachmentIds?: Array<number>;
    attachments?: Array<AttachmentDto>;
}