// src/app/core/api/generated/model/recipient-survey-question-dto.ts
/**
 * B12ForK12 V3 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecipientSurveyChoiceDto } from './recipient-survey-choice-dto';
import { RecipientSurveyExistingAnswerDto } from './recipient-survey-existing-answer-dto';
import { RecipientSurveyConditionDto } from './recipient-survey-condition-dto';


export interface RecipientSurveyQuestionDto { 
    questionInstanceId: string;
    templateQuestionId: string;
    questionOriginId: string;
    order: number;
    text?: string;
    type: number;
    isMandatory: boolean;
    ratingMax?: number;
    maxAttachments?: number;
    contextDisplay?: string;
    contextDetails?: Array<string>;
    choices?: Array<RecipientSurveyChoiceDto>;
    conditions?: Array<RecipientSurveyConditionDto>;
    existingAnswer?: RecipientSurveyExistingAnswerDto;
}