// src/app/core/api/generated/model/recipient-survey-section-dto.ts
/**
 * B12ForK12 V3 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecipientSurveyQuestionDto } from './recipient-survey-question-dto';


export interface RecipientSurveySectionDto { 
    sectionId: string;
    order: number;
    topicType: number;
    topicLabel?: string;
    questions?: Array<RecipientSurveyQuestionDto>;
}