// src/app/core/api/generated/model/submit-survey-response-dto.ts
/**
 * B12ForK12 V3 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SubmitSurveyAnswerDto } from './submit-survey-answer-dto';


export interface SubmitSurveyResponseDto { 
    distributionId: string;
    answers?: Array<SubmitSurveyAnswerDto>;
}