// src/app/core/api/generated/model/survey-session-bootstrap-dto.ts
/**
 * B12ForK12 V3 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecipientSurveyDistributionDto } from './recipient-survey-distribution-dto';


export interface SurveySessionBootstrapDto { 
    surveySessionToken?: string;
    distributionId: string;
    surveyId: string;
    targetUserId: number;
    sessionCreatedAtUtc: string;
    sessionExpiresAtUtc: string;
    submissionAllowed: boolean;
    isSubmitted: boolean;
    distribution?: RecipientSurveyDistributionDto;
}