// src/app/core/api/generated/model/attachment-model-dto.ts
/**
 * B12ForK12 V3 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AttachmentUsersAccessesDTO } from './attachment-users-accesses-dto';


export interface AttachmentModelDTO { 
    id?: number;
    orginId?: number;
    fileName?: string;
    fileUploadName?: string;
    fileSizeInBytes: number;
    fileURL?: string;
    fileThumbnailURL?: string;
    fileContentType?: string;
    fileExtension?: string;
    fileType: number;
    originalFileURL?: string;
    shortenURL?: string;
    attachmentUsersAccesses?: Array<AttachmentUsersAccessesDTO>;
}